Add permissions-based visibility and enforcement for file actions in File Manager.

+++ resources/views/components/file-actions.blade.php
@@ -2,6 +2,7 @@
     'item',
     'isFolder' => false,
     'size' => 'sm',
+    'permissions' => [],
 ])
 
 @php
@@ -20,28 +21,34 @@
     >
         <x-filament::icon icon="heroicon-m-eye" :class="$iconClass" />
     </button>
+    @if ($permissions['canDownload'] ?? true)
+        <button
+            wire:click="downloadFile(@js($item->path))"
+            type="button"
+            class="{{ $buttonClass }} text-gray-400 transition hover:bg-gray-100 hover:text-gray-600 dark:hover:bg-white/10 dark:hover:text-gray-300"
+            title="{{ __('filament-file-manager::file-manager.actions.download') }}"
+        >
+            <x-filament::icon icon="heroicon-m-arrow-down-tray" :class="$iconClass" />
+        </button>
+    @endif
+@endif
+@if ($permissions['canRename'] ?? true)
     <button
-        wire:click="downloadFile(@js($item->path))"
+        wire:click="mountAction('rename', { path: @js($item->path) })"
         type="button"
         class="{{ $buttonClass }} text-gray-400 transition hover:bg-gray-100 hover:text-gray-600 dark:hover:bg-white/10 dark:hover:text-gray-300"
-        title="{{ __('filament-file-manager::file-manager.actions.download') }}"
+        title="{{ __('filament-file-manager::file-manager.actions.rename') }}"
     >
-        <x-filament::icon icon="heroicon-m-arrow-down-tray" :class="$iconClass" />
+        <x-filament::icon icon="heroicon-m-pencil" :class="$iconClass" />
     </button>
 @endif
-<button
-    wire:click="mountAction('rename', { path: @js($item->path) })"
-    type="button"
-    class="{{ $buttonClass }} text-gray-400 transition hover:bg-gray-100 hover:text-gray-600 dark:hover:bg-white/10 dark:hover:text-gray-300"
-    title="{{ __('filament-file-manager::file-manager.actions.rename') }}"
->
-    <x-filament::icon icon="heroicon-m-pencil" :class="$iconClass" />
-</button>
-<button
-    wire:click="mountAction('deleteItem', { path: @js($item->path) })"
-    type="button"
-    class="{{ $buttonClass }} text-gray-400 transition hover:bg-red-50 hover:text-red-500 dark:hover:bg-red-500/10"
-    title="{{ __('filament-file-manager::file-manager.actions.delete') }}"
->
-    <x-filament::icon icon="heroicon-m-trash" :class="$iconClass" />
-</button>
+@if ($permissions['canDelete'] ?? true)
+    <button
+        wire:click="mountAction('deleteItem', { path: @js($item->path) })"
+        type="button"
+        class="{{ $buttonClass }} text-gray-400 transition hover:bg-red-50 hover:text-red-500 dark:hover:bg-red-500/10"
+        title="{{ __('filament-file-manager::file-manager.actions.delete') }}"
+    >
+        <x-filament::icon icon="heroicon-m-trash" :class="$iconClass" />
+    </button>
+@endif
